Add multiplication Factor Property to number line node
This property will later be used in the equality lab tab to add half of
a tick to one side of the number line. Also add an options to
appendtick to choose if it is symmetric or which direction the tick will
appear on a horizontal number line by passing in 'up' or 'down'. I name
it multiplication Factor Property because it multiple the fraction and
number of tick in the equality lab.