update label for bucketNode
The label for the BucketNode is now an HBox that takes an icon and the current fraction node. In IntroModel, we have added an additional derived property called segmentproperty whose value is equal to one over the denominatorproperty value. Modified Bucketnode to take segment property as a parameters. Added an option to beakerNode to choose the width of the tick marks.